% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruction.R
\name{as_instruction}
\alias{as_instruction}
\alias{as_instruction.list}
\title{Converts objects into instruction}
\usage{
as_instruction(x)

\method{as_instruction}{list}(x)
}
\arguments{
\item{x}{an object generated by \code{list()}}
}
\description{
\code{as_instruction} parses a named list or vector as a \code{chewie_instruction}
object
}
\details{
The function helps users and internal functions to deal with object
conversions into \code{chewie_instruction} class. As the package currently views
\code{data.frame} type of objects (later transformed into \code{chewie_scheme} objects)
as the main target of user experience, rows should be parsed as instructions
above all else.
}
