% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheme.R
\name{as_scheme.data.frame}
\alias{as_scheme.data.frame}
\title{Loads scheme from a dataframe}
\usage{
\method{as_scheme}{data.frame}(x)
}
\description{
Loads scheme from a dataframe
}
\concept{scheme}
\keyword{scheme}
