% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheme.R
\name{as_scheme}
\alias{as_scheme}
\alias{as_scheme.data.frame}
\alias{as_scheme.list}
\title{Loads scheme from a dataframe}
\usage{
as_scheme(x)

\method{as_scheme}{data.frame}(x)

\method{as_scheme}{list}(x)
}
\description{
Loads scheme from a dataframe

Loads scheme from a list
}
\concept{scheme}
\keyword{scheme}
