% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheme.R
\name{chew}
\alias{chew}
\title{Chew instructions from scheme}
\usage{
chew(page, scheme)
}
\arguments{
\item{page}{a \code{xml_document} element}

\item{scheme}{the type of extractor to be utilized}
}
\description{
Loads scheme and parses a page using the instructions contained by it.
}
\keyword{chew}
