% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chewie-cache.r
\name{chewie_setup_cache}
\alias{chewie_setup_cache}
\alias{chewie_unset_cache}
\alias{chewie_get_cache}
\title{Set GEDI Parquet Cache}
\usage{
chewie_setup_cache(
  .dir = chewie_default_dir(),
  renviron = "global",
  quiet = FALSE
)

chewie_unset_cache(renviron = "global")

chewie_get_cache()
}
\arguments{
\item{.dir}{character path to the directory that should contain the cache.}

\item{renviron}{character either 'global', 'local' or path to the directory
containing the \code{.Renviron} file to set the \code{CHEWIE_PARQUET_CACHE}
environment and create the directory. If missing, the default cache
directory is used.}
}
\description{
Set GEDI Parquet Cache

Unset GEDI Cache

Get GEDI Cache
}
\details{
This \code{chewie_unset_cache} function will remove the
\code{CHEWIE_PARQUET_CACHE} environment variable from the \code{.Renviron} file.
}
\concept{manage cache}
