% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chewie-find.r
\name{print.chewie.find}
\alias{print.chewie.find}
\title{print a \code{chewie.find} object}
\usage{
\method{print}{chewie.find}(x, ...)
}
\arguments{
\item{x}{chewie.find object to print}

\item{...}{arguments passed to
\code{\link[data.table:print.data.table]{data.table::print.data.table}}}
}
\description{
\code{print.chewie.find} is a method for printing \code{chewie.find}
objects. Essentially a wrapper for \code{print.data.table} with some custom
options. See details for options.
}
\details{
Edit the following options to change the default printing behaviour
of \code{chewie.find} objects: \code{chewie.print.class}, \code{chewie.print.keys},
\code{chewie.print.topn}, \code{chewie.print.nrows}, \code{chewie.print.trunc.cols},
\code{chewie.prettyprint.char}, \code{chewie.print.width}. Additonal options for
\code{print.data.table} can also be set to alter the print layout.
}
\concept{print}
