% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chewie-bbox.r
\name{chewie_bbox}
\alias{chewie_bbox}
\alias{chewie_bbox.SpatRaster}
\alias{chewie_bbox.SpatVector}
\alias{chewie_bbox.sf}
\alias{chewie_bbox.sfc}
\alias{chewie_bbox.stars}
\alias{chewie_bbox.stars_proxy}
\alias{chewie_bbox.numerc}
\title{Get spatial extent}
\usage{
chewie_bbox(x, ...)

\method{chewie_bbox}{SpatRaster}(x, ...)

\method{chewie_bbox}{SpatVector}(x, ...)

\method{chewie_bbox}{sf}(x, ...)

\method{chewie_bbox}{sfc}(x, ...)

\method{chewie_bbox}{stars}(x, ...)

\method{chewie_bbox}{stars_proxy}(x, ...)

\method{chewie_bbox}{numerc}(x, ...)
}
\arguments{
\item{x}{A spatial object, file path or source}

\item{...}{Not used}
}
\value{
A numeric vector of length 4. Values are returned as:
"xmin", "xmax", "ymin", "ymax"
}
\description{
A class agnostic function to return the bounding extent (i.e. bounding box)
of a spatial object.
}
\details{
where x is numeric it should be avector of length for with
cordinates orderd as xmin, ymin, xmax, ymax.
}
\examples{
f <- system.file("ex/elev.tif", package = "terra")
chewie_bbox(f)
chewie_bbox(terra::rast(f))
f2 <- system.file("ex/lux.shp", package = "terra")
chewie_bbox(f2)
}
\concept{spatial helpers (class agnostic)}
