% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chewie-grab.r
\name{chewie_download}
\alias{chewie_download}
\title{Download GEDI data}
\usage{
chewie_download(
  x,
  progress = TRUE,
  timeout = Inf,
  .dir = getOption("chewie.session.cache")
)
}
\arguments{
\item{x}{A chewie.find.x object.}
}
\description{
Download GEDI data from the NASA Earthdata in hdf5 format.
}
