% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chewie-find.r
\name{chewie_find}
\alias{chewie_find}
\title{Find GEDI swaths}
\usage{
chewie_find(
  x,
  gedi_product = c("1B", "2A", "2B"),
  gedi_version = c("v2", "v1"),
  date_range = NULL
)
}
\arguments{
\item{x}{object of class \code{sf}, \code{spatVector}, \code{spatRaster}, \code{sfc}, \code{stars},
\code{stars_proxy}, or \code{numeric} see details.}

\item{gedi_product}{character of GEDI product to search for.}

\item{gedi_version}{character of GEDI version to search for.}

\item{date_range}{character vector of length 2, with start and end dates
in the format "YYYY-MM-DD".}
}
\description{
Find GEDI swaths that intersect with a given spatial object.
}
\details{
Where x is a numeric it must be of length 4 with values corresponding to the
bounding box coordinates in the order xmin, ymin, xmax, ymax.
}
