% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chewie-grab.r
\name{chewie_grab}
\alias{chewie_grab}
\title{Download GEDI data}
\usage{
chewie_grab(x, progress = TRUE, timeout = 7200)
}
\arguments{
\item{x}{A chewie.find.x object.}

\item{progress}{A logical indicating whether to show a progress bar.}

\item{timeout}{A numeric indicating the timeout in seconds.}
}
\description{
Download GEDI data from the NASA Earthdata in hdf5 format.
}
\details{
This function is the main handler for gedi data - it checks the cache to see
if the required GEDI data are already downloaded, and if not, downloads them
from the NASA Earthdata. Once downloaded each file is converted to the
parquet format and saved in the cache directory. This saves a huge amount of
disk space and enables dynamic reading and filtering of the returned "open"
arrow dataset.
}
