% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chewie-health-check.r
\name{chewie_health_check}
\alias{chewie_health_check}
\title{Check the status of the {chewie}}
\usage{
chewie_health_check(.test = TRUE)
}
\arguments{
\item{.test}{logical; whether to test the credentials (requires internet
connection)}
}
\description{
checks the NASA Earthdata Credentials and GEDI Cache
environment variables
}
