% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruction.R
\name{execute_instruction}
\alias{execute_instruction}
\title{Runs an instruction on a HTML page}
\usage{
execute_instruction(page, instruction)
}
\arguments{
\item{page}{a \code{xml_document} object}

\item{instruction}{a \code{chewie_instruction} object}
}
\value{
an \code{chewie_instruction} object containing the extracted element from
the \code{path}/\code{alternative_path} provided.
}
\description{
\code{execute_instruction} executes a \code{chewie_instruction} on a parsed HTML
page and deliver a polished scraped item to the user.
}
