% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{extract_numeric}
\alias{extract_numeric}
\title{Extract texts from \code{xml_node} object.}
\usage{
extract_numeric(field, pattern = NULL)
}
\arguments{
\item{field}{a \code{xml_node} or \code{xml_nodeset} element}

\item{pattern}{a valid RegEx pattern}
}
\value{

}
\description{
Wrapper around \code{extract_text}. Converts extracted texts to numeric.
}
\seealso{
Other extractors: 
\code{\link{extract_price}()},
\code{\link{extract_text}()}
}
\concept{extractors}
\keyword{extractor}
