% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{extract_price}
\alias{extract_price}
\title{Extract prices from \code{xml_node} elements.}
\usage{
extract_price(field, pattern = NULL)
}
\arguments{
\item{field}{a \code{xml_node} element}

\item{pattern}{a valid RegEx pattern}
}
\description{
Wrapper around \code{extract_text}. Extracts digits from a string, convert them
into a single integer and divide it by 100, considering that prices won't have
more than 2 decimal digits.
}
\seealso{
Other extractors: 
\code{\link{extract_numeric}()},
\code{\link{extract_text}()}
}
\concept{extractors}
\keyword{extractor}
