% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{extract_text}
\alias{extract_text}
\title{Extract texts from \code{xml_document} or \code{xml_node} element.}
\usage{
extract_text(field, pattern = NULL)
}
\arguments{
\item{field}{a \code{xml_document} or \code{xml_node} element}

\item{pattern}{a valid RegEx pattern}
}
\description{
Text parser using methods from both \code{stringr} and \code{rvest}. Automatically
vectorised by both packages properties, can return a list of text or a
single text
}
\seealso{
Other extractors: 
\code{\link{extract_numeric}()},
\code{\link{extract_price}()}
}
\concept{extractors}
\keyword{extractor}
