% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chewie-waveforms.r
\name{extract_waveforms}
\alias{extract_waveforms}
\title{Collect waveform data from GEDI 1B data}
\usage{
extract_waveforms(x)
}
\arguments{
\item{x}{A tibble or data.frame object returned from \code{collect_gedi}.}
}
\value{
a tibble containing the waveform data.
}
\description{
Collect waveform data from GEDI 1B data
}
\details{
This function is used to collect the waveform data from GEDI 1B data returned
from \code{collect_gedi}. The waveform amplitude is converted to a float value with
a precision of 0.0001 (i.e. 4 decimal places) and the elevation is calculated
from the \code{elevation_bin0}, \code{elevation_lastbin} and \code{rx_sample_count} columns.

This function converts the waveform lists for each footprint to long form and
merges all shots into a single tibble. The advantage of this is that it
enables the analysis and comparison of multiple plots. If you wish to filter
the shots that you want to analyse, simply filter the results of 'collect_gedi'
before passing to this function.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
prairie_creek <- sf::read_sf(
  system.file("geojson", "prairie-creek.geojson", package = "chewie")
)
prairie_creek_find_1b <- find_gedi(prairie_creek,
  gedi_product = "1B",
  date_start = "2022-01-01", date_end = "2022-04-01",
  cache = FALSE
)

prairie_creek_grab_1b <- grab_gedi(prairie_creek_find_1b)

prairie_creek_1b_sf <- collect_gedi(
  prairie_creek_grab_1b,
  prairie_creek_find_1b
)

prairie_creek_waveforms <- collect_waveforms(prairie_creek_1b_sf)
print(prairie_creek_waveforms)
\dontshow{\}) # examplesIf}
}
