% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruction.R
\name{find_elements}
\alias{find_elements}
\title{Find elements in a parsed HTML page}
\usage{
find_elements(page, instruction)
}
\arguments{
\item{page}{a \code{xml_document} object}

\item{instruction}{a \code{chewie_instruction} object}
}
\value{
correspondent \code{xml_node} or \code{xml_nodeset} of the described \code{path}/
\code{alternative_path}
}
\description{
\code{find_elements} scrapes a HTML page using \code{rvest::html_elements} and
paths described in a \code{chewie_instruction} object. If \code{path} argument in
instruction doesn't find any results it will automatically switch to an
\code{alternative_path} if it is described in instruction as well.
}
\examples{
\dontrun{
sample_instruction <- instruction(
  title = "price_header",
  path = "h1:nth-of-type(1)",
  selector = "css",
  alternative_path = "h2:nth-of-type(2)",
  parse_as = "text"
)

sample_page <- rvest::read_html(response)

results <- find_elements(sample_page, sample_instruction)
}

}
