% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.r
\name{find_gedi_product}
\alias{find_gedi_product}
\title{get the GEDI product from a \code{chewie.find} object}
\usage{
find_gedi_product(x)
}
\description{
get the GEDI product from a \code{chewie.find} object
}
