% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruction.R
\name{instruction}
\alias{instruction}
\alias{instruction_set_result}
\title{Base instruction class}
\usage{
instruction(
  title,
  path,
  selector = "css",
  alternative_path = NULL,
  parse_as = NULL,
  pattern = NULL
)

instruction_set_result(x, result)
}
\arguments{
\item{title}{an arbitrary name to the scraped object}

\item{path}{a css or xpath path to the object to be scraped}

\item{selector}{whether \code{path}/\code{alternative_path} is a css or xpath selector,
defaults to \code{NULL}}

\item{alternative_path}{an alternative css or xpath path to the object to be
scraped}

\item{parse_as}{indicates if an extractor should be applied to the resulting
scraped item. Currently available extractors are:
\itemize{
\item \code{extract_text}
\item \code{extract_numeric}
\item \code{extract_table}
\item \code{extract_date}
\item \code{extract_datetime}
\item \code{extract_timedelta}
\item \code{extract_price}.
}}

\item{pattern}{a RegEx pattern to be applied before parsing}
}
\value{
a \code{chewie_instruction} object
}
\description{
\code{instruction} provides the guidelines to \code{chewie} scraping by setting all of
searching and returning object parameters.
}
\examples{
sample_instruction <- instruction(
  title = "price_header",
  selector = "css",
  path = "h1:nth-of-type(1)",
  alternative_path = "h2:nth-of-type(2)",
  parse_as = "text"
)

}
