% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{parse_field}
\alias{parse_field}
\title{Extract texts from \code{xml_document} or \code{xml_node} element.}
\usage{
parse_field(field, as = NULL, pattern = NULL)
}
\arguments{
\item{field}{a \code{xml_node} or \code{xml_nodeset} element}

\item{pattern}{a valid pattern pattern}

\item{output_type}{the type of extractor to be utilized}
}
\description{
Text parser using methods from both \code{stringr} and \code{rvest}. Automatically
vectorised by both packages properties, can return a list of text or a
single text
}
\keyword{parser}
