% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chewie-find.r
\name{plot.chewie.find}
\alias{plot.chewie.find}
\title{print a \code{chewie.find} object}
\usage{
\method{plot}{chewie.find}(x, swath_col, land_col = "#63a88840", ...)
}
\arguments{
\item{x}{chewie.find object to print}

\item{swath_col}{character; color of the swaths.}

\item{land_col}{character; color of the land.}

\item{...}{arguments passed to \code{print.sf}}
}
\value{
a base R plot.
}
\description{
\code{print.chewie.find} is a method for printing \code{chewie.find}
objects.
}
