% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chewie-find.r
\name{print.chewie.find}
\alias{print.chewie.find}
\title{print a chewie object}
\usage{
\method{print}{chewie.find}(x, ...)
}
\arguments{
\item{x}{chewie object to print}
}
\description{
print a chewie object
}
