% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheme.R
\name{chew}
\alias{chew}
\title{Chew instructions from scheme}
\usage{
chew(scheme, page = NULL, url = NULL, ...)
}
\arguments{
\item{scheme}{a \code{scheme} type object containing scraping instructions}

\item{page}{a product of \code{rvest::read_html()}}

\item{url}{a link to a page}
}
\description{
Loads scheme and parses a page using the instructions contained by it.
}
\keyword{chew}
