% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chewie-cache.r
\name{chewie_setup_cache}
\alias{chewie_setup_cache}
\alias{chewie_unset_cache}
\alias{chewie_get_cache}
\alias{chewie_clear_find_cache}
\alias{chewie_clear_h5_temp_cache}
\alias{chewie_clear_parquet_cache}
\title{Set GEDI Parquet Cache}
\usage{
chewie_setup_cache(
  .dir = chewie_default_dir(),
  renviron = "user",
  quiet = FALSE
)

chewie_unset_cache(renviron = "user")

chewie_get_cache()

chewie_clear_find_cache()

chewie_clear_h5_temp_cache()

chewie_clear_parquet_cache(directories = c("none", "1B", "2A", "2B", "4A"))
}
\arguments{
\item{.dir}{character path to the directory that should contain the cache.}

\item{renviron}{character either 'user', '"project"' or path to the directory
containing the \code{.Renviron} file to set the \code{CHEWIE_PARQUET_CACHE}
environment and create the directory. If missing, the default cache
directory is used.}

\item{quiet}{logical if \code{TRUE} then no messages are printed.}

\item{directories}{character; the GEDI product parquet directories to clear.}
}
\description{
Get the path to the GEDI cache directory.
}
\details{
This \code{chewie_unset_cache} function will remove the
\code{CHEWIE_PARQUET_CACHE} environment variable from the \code{.Renviron} file.

\code{chewie_clear_find_cache} deletes the cached .rds files in the GEDI find
cache directory, located in \code{getOption("chewie.find.gedi.cache")}.

\code{chewie_clear_h5_temp_cache} deletes the cached .h5 files in the GEDI h5 -
these files sometimes persist when a download in incomplete or there has
been an error with the download. If you are having trouble downloading data,
running this command could well help.

\code{chewie_clear_parquet_cache} deletes the cached .parquet files in the GEDI
parquet cache directory, located in \code{getOption("chewie.parquet.cache")}.

This function should almost never be used. Situations where this may be
useful might include - requiring to free up disk space, if you are
experiencing issues with the cache, or if there are package updates which
render the existing cache irrelevant or incompatible.
}
\concept{manage cache}
