% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chewie-creds.r, R/utils-envs.r
\name{chewie_creds}
\alias{chewie_creds}
\alias{chewie_env_clean}
\alias{chewie_set_env}
\alias{chewie_get_env}
\title{Set NASA Earthdata Credentials}
\usage{
chewie_creds(
  .netrc = NULL,
  .path = NULL,
  .force = FALSE,
  .usr = NULL,
  .pwd = NULL,
  .quiet = FALSE,
  .test = TRUE
)

chewie_env_clean(renviron = "global", .check = interactive())

chewie_set_env(.netrc, renviron = "global")

chewie_get_env(.env = "CHEWIE_NETRC")
}
\arguments{
\item{.netrc}{character path to an existing \code{.netrc} file.}

\item{.path}{character path to set where to create the \code{.netrc} file}

\item{.force}{logical whether to overwrite an existing \code{.netrc} or
\code{CHEWIE_NETRC} environment variable.}

\item{.usr}{character NASA Earthdata username.}

\item{.pwd}{character NASA Earthdata password.}

\item{.quiet}{logical whether to suppress the registration prompt and
messaging.}

\item{renviron}{character either 'global', 'local' or path to the directory
containing the \code{.Renviron} file to set the \code{CHEWIE_NETRC} environment.}
}
\description{
Set NASA Earthdata credentials for use in \code{chewie} functions.
}
\details{
The NASA Earthdata API requires a username and password to access data. This
is provided via the config settings in the form of a \code{.netrc} file. The
\code{chewie_creds} function can help to generate this file and set its location
as an environment variable for use in \code{chewie} functions. In theory this
should only need to be done once, but if you need to change your credentials
you can use the \code{force} argument to overwrite the existing \code{.netrc} file.

\code{chewie_env_clean} can be used to manually remove the \code{CHEWIE_NETRC}
environment and delete the associated \code{.netrc} file.

\code{chewie_set_env} is most likely not required but can be used to
manually set the \code{CHEWIE_NETRC} environment variable which is used for
authenticating downloads from the NASA Earthdata API.

\code{chewie_get_env} can be used to manually get the \code{CHEWIE_NETRC}
environment, providing the file path to the \code{.netrc} file.
}
\concept{manage credentials}
\concept{manage credentials
return character file path for \code{.netrc} file}
