% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chewie-creds.r
\name{chewie_creds}
\alias{chewie_creds}
\alias{chewie_clean_netrc}
\alias{chewie_set_netrc_env}
\alias{chewie_get_netrc}
\title{Set NASA Earthdata Credentials}
\usage{
chewie_creds(
  netrc = NULL,
  path = NULL,
  force = FALSE,
  usr = NULL,
  pwd = NULL,
  quiet = FALSE,
  test = TRUE,
  renviron = "user"
)

chewie_clean_netrc(renviron = "user")

chewie_set_netrc_env(netrc, renviron = "user")

chewie_get_netrc()
}
\arguments{
\item{netrc}{character path to an existing \code{.netrc} file.}

\item{path}{character path to set where to create the \code{.netrc} file}

\item{force}{logical whether to overwrite an existing \code{.netrc} or
\code{CHEWIE_NETRC} environment variable.}

\item{usr}{character NASA Earthdata username.}

\item{pwd}{character NASA Earthdata password.}

\item{quiet}{logical whether to suppress the registration prompt and
messaging.}

\item{test}{logical whether to test the credentials after setting them.}

\item{renviron}{character either 'user', '"project"' or path to the directory
containing the \code{.Renviron} file to set the \code{CHEWIE_NETRC} environment.}
}
\description{
Set NASA Earthdata credentials for use in \code{chewie} functions.
}
\details{
The NASA Earthdata API requires a username and password to access data. This
is provided via the config settings in the form of a \code{.netrc} file. The
\code{chewie_creds} function can help to generate this file and set its location
as an environment variable for use in \code{chewie} functions. In theory this
should only need to be done once, but if you need to change your credentials
you can use the \code{force} argument to overwrite the existing \code{.netrc} file.

\code{chewie_env_clean} can be used to manually remove the \code{CHEWIE_NETRC}
environment and delete the associated \code{.netrc} file.

\code{chewie_set_netrc_env} is most likely not required but can be used to
manually set the \code{CHEWIE_NETRC} environment variable which is used for
authenticating downloads from the NASA Earthdata API.

\code{chewie_get_netrc} can be used to manually get the \code{CHEWIE_NETRC}
environment, providing the file path to the \code{.netrc} file used to
authenticate any requests to the NASA Earthdata API
}
\concept{manage credentials}
\concept{manage credentials
return character file path for \code{.netrc} file}
