% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chewie-find.r
\name{find_gedi}
\alias{find_gedi}
\alias{chewie_clear_find_cache}
\title{Find GEDI swaths}
\usage{
find_gedi(
  x,
  gedi_product = c("1B", "2A", "2B"),
  date_start = NULL,
  date_end = NULL,
  intersects = TRUE,
  cache = TRUE
)

chewie_clear_find_cache()
}
\arguments{
\item{x}{object of class \code{sf}, \code{spatVector}, \code{spatRaster}, \code{sfc}, \code{stars},
\code{stars_proxy}, or \code{numeric} see details.}

\item{gedi_product}{character of GEDI product to search for.}

\item{date_start}{character or \code{POSIXct} of the start date to search for GEDI
data. If \code{NULL} defaults to the start of GEDI operations (2019-03-25).}

\item{date_end}{character or \code{POSIXct} of the end date to search for GEDI
data. If \code{NULL} defaults to the current date.}

\item{intersects}{logical indicating whether to return only swaths that
intersect with the given spatial object. If FALSE all swaths that are
contained within the bounding box of the spatial object are returned.}

\item{cache}{logical indicating whether to cache the results of the GEDI
search. If TRUE the results of the search will be cached in the directory
defined by: \code{getOption("chewie.find.gedi.cache")}. If FALSE the results will
not be cached.}

\item{gedi_version}{character of GEDI version to search for.}
}
\description{
Find GEDI swaths that intersect with a given spatial object.
}
\details{
Where x is a numeric it must be of length 4 with values corresponding to the
bounding box coordinates in the order xmin, ymin, xmax, ymax.

\code{chewie_clear_find_cache} deletes the cached .rds files in the GEDI find
cache directory, located in \code{getOption("chewie.find.gedi.cache")}.
}
\concept{find GEDI}
\concept{find GEDI.}
