% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chewie-find.r
\name{chewie_find}
\alias{chewie_find}
\title{Find GEDI swaths}
\usage{
chewie_find(
  x,
  gedi_product = c("1B", "2A", "2B"),
  gedi_version = c("v2", "v1"),
  date_start = NULL,
  date_end = NULL
)
}
\arguments{
\item{x}{object of class \code{sf}, \code{spatVector}, \code{spatRaster}, \code{sfc}, \code{stars},
\code{stars_proxy}, or \code{numeric} see details.}

\item{gedi_product}{character of GEDI product to search for.}

\item{gedi_version}{character of GEDI version to search for.}

\item{date_start}{character or \code{POSIXct} of the start date to search for GEDI
data. If \code{NULL} defaults to the start of GEDI operations (2019-03-25).}

\item{date_end}{character or \code{POSIXct} of the end date to search for GEDI
data. If \code{NULL} defaults to the current date.}
}
\description{
Find GEDI swaths that intersect with a given spatial object.
}
\details{
Where x is a numeric it must be of length 4 with values corresponding to the
bounding box coordinates in the order xmin, ymin, xmax, ymax.
}
