% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chewie-show.r
\name{chewie_show}
\alias{chewie_show}
\title{Show an interactive map of a chewie_* object.}
\usage{
chewie_show(x, ...)
}
\arguments{
\item{x}{object of class \code{chewie_find} or \code{chewie_grab}.}
}
\value{
interactive leaflet map
}
\description{
Show an interactive map of a chewie_* object.
}
