% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chewie-show.r
\name{chewie_show}
\alias{chewie_show}
\title{Show an interactive map of a chewie_* object.}
\usage{
chewie_show(x, ...)
}
\arguments{
\item{x}{object of class \code{chewie_find} or \code{chewie_grab}.}

\item{...}{additional arguments passed to \code{mapview::mapview}.}

\item{time_group}{character; if \code{interactive = TRUE}, group the swaths by
this time unit. Must be one of \code{"year"} or \code{"month"}.}

\item{alpha}{numeric; transparency of the swaths.}

\item{swath_pal}{character; name of the color palette to use for the swaths.
see \code{hcl.pals()} for a list of available palettes.}

\item{aoi_color}{character; color of the AOI polygon.}

\item{zoom}{numeric; zoom level of the map. If \code{NULL} (default), the zoom
level is determined automatically by the extent of the GEDI swaths.}

\item{interactive}{logical; if \code{TRUE} (default), return an interactive map.
If \code{FALSE}, return a static map.}

\item{file}{character; if \code{interactive = FALSE}, save the static map to this
file.}
}
\value{
interactive leaflet map
}
\description{
Show an interactive map of a chewie_* object.
}
\details{
This function is a wrapper around \code{mapview::mapview()}. It is designed to
work with \code{chewie_find()} and \code{chewie_grab()} objects.
}
