% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chewie-arrow.r
\name{collect_gedi}
\alias{collect_gedi}
\title{Collect GEDI data}
\usage{
collect_gedi(x, gedi_find)
}
\arguments{
\item{x}{An arrow dataset object.}

\item{gedi_find}{The chewie.find object used to obtain \code{x}.}
}
\value{
an sf object
}
\description{
Collect GEDI data, returned from \code{grab_gedi}, as an sf object.
}
\details{
This function is used to collect the GEDI data returned from \code{grab_gedi} as
an sf object. It is largely a wrapper for dplyr::collect but converts to sf
and filters the gedi footprints based on the search extent attributed to the
\code{chewie.find} object.
It is strongly recomended that you make the most of the ability to to edit
the gedi data on read by using the \code{dplyr} verbs before collecting the data.
This will save a lot of time and memory. However, make sure that, when
selecting columns, you do not remove the \code{lat_lowestmode} and
\code{lon_lowestmode} columns as these are required to create the geometry column.
}
