% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{extract_datetime}
\alias{extract_datetime}
\title{Extract datetime from \code{xml_node} elements.}
\usage{
extract_datetime(field, pattern = NULL, ...)
}
\arguments{
\item{field}{a \code{xml_node} or \code{xml_nodeset} element}

\item{pattern}{a valid RegEx pattern}
}
\description{
Extract datetime from \code{xml_node} elements.
}
