% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chewie-find.r
\name{plot.chewie.find}
\alias{plot.chewie.find}
\title{plot a \code{chewie.find} object}
\usage{
\method{plot}{chewie.find}(x, swath_col = "#903ca586", aoi_col = "#cecece", ...)
}
\arguments{
\item{x}{chewie.find object to print}

\item{swath_col}{character; color of the swaths.}

\item{aoi_col}{character; color of the aoi.}

\item{...}{arguments passed to \code{print.sf}}
}
\value{
a base R plot.
}
\description{
\code{plot.chewie.find} is a method for plotting \code{chewie.find}
objects.
}
